requirements.txt
numpy
scipy
soundfile
torch
tqdm
pyyaml
pandas
transformers
matplotlib
librosa
google-genai
panns-inference
